Support refreshing location when weather data is null
It appears that sometimes a BOM observatory is undergoing something and
the result of it is the weather data suddenly become absent. When this
happens, The location search from BOM website will re-locate the nearest
observatory to the other one.

This commit supports refreshing the location (thus re-locating weather
source URL) when weather appears to be null.