fix Linux (gnome 3) behaviors
Actually commit d49c6513299885b5b3f077a14f85012361aea9c8 causes a
regression such that a left click on the icon on MacOS won't show the
main window.

A further fix is this commit which adds an additional item in the menu
bar so that the main window can be shown by clicking on the menu item
instead. Although this doesn't make using it on Linux much better, it
does fix the regression.